requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.3
MarkupSafe==0.23
psycopg2==2.6.1
python-dateutil==2.4.2
python-editor==0.5
pytz==2015.6
requests==2.7.0
six==1.9.0
SQLAlchemy==1.0.11
Werkzeug==0.11.3
wheel==0.24.0
